Reset response queue on lost socket connection to redis
In cases where redis-server gets blocked for some reason,
there'll be for sure outstanding replies that haven't arrived
yet. If redis then dies (or most likely, have it failover to a replica)
then the connection is lost, at this point the queue of pending replies
must also be cleared otherwise when the connection is re-established the
state machine will be out of sync and requests will consistently time
out.